requirements.txt
fastapi[all]
requests
python-dotenv
pytz
motor[srv]